decoder: send events for EXISTS and EXPUNGE
The new state class tracks last seen message number, this is essential
for fetching and IDLE because we need to know the last seen message
  number in order to use those commands.

+++ src/main/java/com/hubspot/imap/imap/command/BaseCommand.java
@@ -12,7 +12,7 @@ public class BaseCommand implements Command {
   protected final String tag;
   protected final List<String> args;
 
-  public BaseCommand(CommandType type, int tag, String... args) {
+  public BaseCommand(CommandType type, long tag, String... args) {
     this.type = type;
     this.tag = "A" + String.valueOf(tag);
     this.args = Lists.newArrayList(args);
