+++ src/main/java/com/hubspot/imap/imap/command/ListCommand.java
@@ -3,7 +3,7 @@ package com.hubspot.imap.imap.command;
 import com.hubspot.imap.utils.GmailUtils;
 
 public class ListCommand extends BaseCommand {
-  public ListCommand(int tag, String reference, String query) {
+  public ListCommand(long tag, String reference, String query) {
     super(CommandType.LIST, tag, GmailUtils.quote(reference), GmailUtils.quote(query));
   }
 }
